@@ -3,44 +3,41 @@
   "version": "0.0.1",
   "private": true,
   "scripts": {
-    "dev": "vite dev --host",
+    "dev": "vite dev",
     "build": "vite build",
     "preview": "vite preview",
-    "test": "npm run test:integration && npm run test:unit",
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-    "lint": "prettier --plugin-search-dir . --check . && eslint .",
-    "format": "prettier --plugin-search-dir . --write .",
-    "test:integration": "playwright test",
-    "test:unit": "vitest"
+    "lint": "eslint ."
   },
   "devDependencies": {
-    "@friendofsvelte/toggle": "^0.0.2",
     "@iconify/json": "^2.2.239",
     "@playwright/test": "^1.28.1",
-    "@sveltejs/adapter-auto": "^2.0.0",
     "@sveltejs/adapter-cloudflare": "^2.3.3",
-    "@sveltejs/kit": "^1.20.4",
+    "@sveltejs/kit": "^2.0.0",
+    "@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
     "@typescript-eslint/eslint-plugin": "^5.45.0",
     "@typescript-eslint/parser": "^5.45.0",
     "autoprefixer": "^10.4.14",
-    "eslint": "^8.28.0",
+    "eslint": "^8.56.0",
     "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-svelte": "^2.30.0",
+    "eslint-plugin-svelte": "^2.36.0-next.4",
     "iconify-icon": "^1.0.8",
     "postcss": "^8.4.24",
     "postcss-load-config": "^4.0.1",
     "prettier": "^2.8.0",
     "prettier-plugin-svelte": "^2.10.1",
-    "svelte": "^4.0.5",
-    "svelte-check": "^3.4.3",
+    "svelte": "^5.0.0",
+    "svelte-check": "^3.6.0",
     "svelte-preprocess": "^5.0.4",
-    "tailwindcss": "^3.3.2",
+    "tailwindcss": "^3.4.9",
     "tslib": "^2.4.1",
     "typescript": "^5.0.0",
-    "unplugin-icons": "^0.19.2",
-    "vite": "^4.4.2",
-    "vitest": "^0.32.2"
+    "vite": "^5.0.3",
+    "unplugin-icons": "^0.19.2"
   },
-  "type": "module"
+  "type": "module",
+  "dependencies": {
+    "@friendofsvelte/toggle": "^0.0.2-svelte.5.docup"
+  }
 }
